+++ CRM/Civirules/Form/RuleAction.php
@@ -130,7 +130,7 @@ class CRM_Civirules_Form_RuleAction extends CRM_Core_Form {
     }
     $actionList = array(' - select - ') + CRM_Civirules_Utils::buildActionList();
     asort($actionList);
-    $attributes = array('class' => 'crm-select2');
+    $attributes = array('class' => 'crm-select2 huge');
     if (empty($this->ruleActionId)) {
       $this->add('select', 'rule_action_select', ts('Select Action'), $actionList, true, $attributes);
     }